fixed --getpkgbuild

+++ src/yaourt.sh
@@ -1421,7 +1421,7 @@ case "$MAJOR" in
 	fi
 	PKG=${args[0]}
 	#msg "Get PKGBUILD for $PKG"
-	if `is_unsupported $PKG`; then
+	if ! `isavailable $PKG` && `is_unsupported $PKG`; then
 		eval $INENGLISH wget "http://aur.archlinux.org/packages/$PKG/$PKG.tar.gz" || { error $(eval_gettext '$PKG not found in AUR.'); die 1; }
 		tar xzf $PKG.tar.gz --transform="s,$PKG,," 2>/dev/null
 		rm $PKG.tar.gz
